requirements.txt
openai==1.41.0
pydantic==2.8.2
python-dotenv==1.0.1